make RenderRTF a reusable component
